Simplify and make Request safe to forget
Created two kinds of Scope objects:

  - LocalScope<'a> to track Requests with arbitrary lifetime 'a
  - StaticScope (no bookkeeping, but only 'static Requests)

The Request trait has been squashed into a single Request<'a, S> data
type, where S is the type of scope that this Request is associated with
and 'a describes the lifetime of this scope.  Any buffers associated
with the request must outlive 'a.

Fixes # 17.