Add libffi and UserOperation
MPI_User_function doesn't have a "user data" pointer so there's no way
to write a single wrapper function that delegates work to arbitrary
closures.  As a result, we have to use libffi to create functions
dynamically.

Completion of all nonblocking requests must outlive the UserOperation.
The fact that the MPI implementation uses reference counting for MPI_Op
won't save us as the closure is destroyed as soon as UserOperation is
dropped.