Expand semantics of EquivalenceFromAnyBytes
EquivalenceFromAnyBytes now requires that the type can be safely and
fully initialized from a properly sized, arbitrary MPI message.

This also fixes a bunch of soundness holes in routines that construct
and return a receive value.

Removes the transmute module and folds into the datatype module now that
it's fairly empty.