MSRV 1.57
We're now getting various transitive dependencies that have updated to
requires 1.56 and 1.57, so as I understand it, we either need to pin
those or acknowledge that MSRV is now 1.57. 1.56 is the first version
that supports the rust-version key, which should help with this
reliability. So long as we're upgrading to this version, we may as well
update our other dependencies.