doc: improve docs for ready send
* Quote from the standard about erroneous use of ready send.

* Add loud warning in `examples/immediate.rs` where ready send is
used (sending to self) since it appears people have copied this example
into uses in which ready send is invalid.

Arguably, ready send should become unsafe because it's so risky to use
and unlikely to be caught by the MPI implementation's run-time checking.
I.e., detection of misuse will be non-deterministic and can easily be
very unlikely in practice.