make ready send unsafe (fix #182)
Ready-mode sends requirements are difficult to ensure without heavy
synchronization (that would outweight the possible performance gains).
The ordering semantics are subject to race conditions and usually not
detectable by the implementation, thus erroneous use may not produce an
MPI error as would be possible for a quality implementation of most
other MPI routines.